"gitgnore, renaming packages, correction PR2"
