--- src/main/java/com/pmm/budgetcalculator/expensing/repositories/ExpenseType.java
@@ -1,22 +1,0 @@
-package com.pmm.budgetcalculator.expensing.repositories;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-
-@Entity
-@Data
-@EqualsAndHashCode(callSuper = false)
-@NoArgsConstructor
-@AllArgsConstructor
-public class ExpenseType extends BaseEntity {
-
-    String expenseType;
-    @ManyToOne(fetch = FetchType.EAGER)
-    private Expense expense;
-
-}
