+++ src/main/java/com/pmm/budgetcalculator/expensing/repository/ExpenseTypeRepository.java
@@ -1,8 +1,7 @@
-package com.pmm.budgetcalculator.expensing.repositories;
+package com.pmm.budgetcalculator.expensing.repository;
 
+import com.pmm.budgetcalculator.expensing.entity.ExpenseType;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 import java.util.Optional;
